Changing from long to uint32_t
- Is important to be sure that those variables has 32 bits

+++ src/CS5490.cpp
@@ -51,7 +51,7 @@ void CS5490::begin(int baudRate){
 /******* Write a register by the serial communication *******/
 /* data bytes pass by data variable from this class */
 
-void CS5490::write(int page, int address, long value){
+void CS5490::write(int page, int address, uint32_t value){
 
 	uint8_t checksum = 0;
 	for(int i=0; i<3; i++)
@@ -404,7 +404,7 @@ double CS5490::getTime(){
 /*              PUBLIC METHODS - Read Register                */
 /**************************************************************/
 
-long CS5490::readReg(int page, int address){
+uint32_t CS5490::readReg(int page, int address){
 	uint32_t value = 0;
 	this->read(page, address);
 	value = value + data[2] << 8;
